move trace protobuf definitions into their own crate
