+++ trace-normalization/src/lib.rs
@@ -5,10 +5,6 @@
 
 #![deny(clippy::all)]
 
-pub mod pb {
-    include!("./pb/pb.rs");
-}
-
 pub mod normalizer;
 
 pub mod normalize_utils;
