+++ trace-obfuscation/Cargo.toml
@@ -5,9 +5,9 @@ authors = ["David Lee <david.lee@datadoghq.com>"]
 edition = "2021"
 
 [dependencies]
-prost = "0.11.6"
 anyhow = "1.0"
 regex = "1"
+datadog-trace-protobuf = { path = "../trace-protobuf" }
 
 [dev-dependencies]
 duplicate = "0.4.1"
